2010-10-13  Sancho Lerena <slerena@artica.es>
    * Core.pm: Force XML::Simple to use XML::Parser instead SAX to manage XML
    due a bug processing some XML with blank spaces.

    * DataServer.pm: Checking for BADXML is 2 times with 2 secs of delay now. 
    Aditional check before deleting / procesing XML.

    * pandora_server: Added handle of BadXML error to avoid see "unhandled" 
    errors on log, and set to verbose level 2.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@3388 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_server/lib/PandoraFMS/DataServer.pm
@@ -145,10 +145,10 @@ sub data_consumer ($$) {
 	# Double check that the file exists
 	return unless (-f $file_name);
 
-	# Try to parse the XML 3 times
+	# Try to parse the XML 2 times, with a delay between tries of 2 seconds
 	my $xml_data;
 
-	for (1..3) {
+	for (0..1) {
 		eval {
 			threads->yield;
 			$xml_data = XMLin ($file_name, forcearray => 'module');
@@ -157,12 +157,15 @@ sub data_consumer ($$) {
 		# Invalid XML
 		if ($@) {
 			$xml_err = $@;
-			sleep (1);
+			sleep (2);
 			next;
 		}
 
 		# Ignore the timestamp in the XML and use the file timestamp instead
 		$xml_data->{'timestamp'} = strftime ("%Y-%m-%d %H:%M:%S", localtime((stat($file_name))[9])) if ($pa_config->{'use_xml_timestamp'} eq '1' || ! defined ($xml_data->{'timestamp'}));
+
+    	# Double check that the file exists
+    	return unless (-f $file_name);
 
 		unlink ($file_name);
 		process_xml_data ($self->getConfig (), $file_name, $xml_data, $self->getServerID (), $self->getDBH ());
